Add groups model
Groups are where we will center our member count info, rather than leaving its count hard coded on the helper.

We are leaving the api open in an insecure way but left a TODO to refactor this approach.

We are also improving our seeds to create groups when setting up the app.